requirements.txt
pynacl
rdflib
pytest
pytest-json-report
